rename HBT freezer to collation (rename files)
