+++ silkworm/db/blocks/headers/header_segment_collation.cpp
@@ -14,52 +14,38 @@
    limitations under the License.
 */
 
-#include "txn_snapshot_freezer.hpp"
+#include "header_segment_collation.hpp"
 
 #include <stdexcept>
 
 #include <silkworm/db/access_layer.hpp>
 #include <silkworm/infra/common/log.hpp>
 
-#include "txn_segment.hpp"
+#include "header_segment.hpp"
 
 namespace silkworm::db {
 
-void TransactionSnapshotFreezer::copy(ROTxn& txn, const FreezerCommand& command, snapshots::SegmentFileWriter& file_writer) const {
+void HeaderSnapshotFreezer::copy(ROTxn& txn, const FreezerCommand& command, snapshots::SegmentFileWriter& file_writer) const {
     BlockNumRange range = command.range;
-    snapshots::TransactionSegmentWriter writer{file_writer};
+    snapshots::HeaderSegmentWriter writer{file_writer};
     auto out = writer.out();
-    auto system_tx = snapshots::empty_system_tx();
-
     for (BlockNum i = range.start; i < range.end; ++i) {
-        BlockBody body;
-        bool found = read_canonical_body(txn, i, /* read_senders = */ true, body);
-        if (!found) throw std::runtime_error{"TransactionSnapshotFreezer::copy missing body for block " + std::to_string(i)};
-
-        *out++ = system_tx;
-        for (auto& value : body.transactions) {
-            *out++ = value;
-        }
-        *out++ = system_tx;
+        auto value_opt = read_canonical_header(txn, i);
+        if (!value_opt) throw std::runtime_error{"HeaderSnapshotFreezer::copy missing header for block " + std::to_string(i)};
+        *out++ = *value_opt;
     }
 }
 
-void TransactionSnapshotFreezer::cleanup(RWTxn& txn, BlockNumRange range) const {
+void HeaderSnapshotFreezer::cleanup(RWTxn& txn, BlockNumRange range) const {
     for (BlockNum i = range.start, count = 1; i < range.end; ++i, ++count) {
         auto hash_opt = read_canonical_header_hash(txn, i);
         if (!hash_opt) continue;
-        auto hash = *hash_opt;
+        auto& hash = *hash_opt;
 
-        delete_senders(txn, hash, i);
-
-        auto body_opt = read_canonical_body_for_storage(txn, i);
-        if (body_opt) {
-            auto& body = *body_opt;
-            delete_transactions(txn, body.base_txn_id + 1, body.txn_count - 2);
-        }
+        delete_header(txn, i, hash);
 
         if ((count > 10000) && ((count % 10000) == 0)) {
-            log::Debug("TransactionSnapshotFreezer") << "cleaned up until block " << i;
+            log::Debug("HeaderSnapshotFreezer") << "cleaned up until block " << i;
         }
     }
 }
