+++ silkworm/db/transactions/txn_segment_collation.cpp
@@ -14,44 +14,52 @@
    limitations under the License.
 */
 
-#include "body_snapshot_freezer.hpp"
+#include "txn_segment_collation.hpp"
 
 #include <stdexcept>
 
 #include <silkworm/db/access_layer.hpp>
 #include <silkworm/infra/common/log.hpp>
 
-#include "body_segment.hpp"
+#include "txn_segment.hpp"
 
 namespace silkworm::db {
 
-void BodySnapshotFreezer::copy(ROTxn& txn, const FreezerCommand& command, snapshots::SegmentFileWriter& file_writer) const {
+void TransactionSnapshotFreezer::copy(ROTxn& txn, const FreezerCommand& command, snapshots::SegmentFileWriter& file_writer) const {
     BlockNumRange range = command.range;
-    uint64_t base_txn_id = command.base_txn_id;
-
-    snapshots::BodySegmentWriter writer{file_writer};
+    snapshots::TransactionSegmentWriter writer{file_writer};
     auto out = writer.out();
+    auto system_tx = snapshots::empty_system_tx();
+
     for (BlockNum i = range.start; i < range.end; ++i) {
-        auto value_opt = read_canonical_body_for_storage(txn, i);
-        if (!value_opt) throw std::runtime_error{"BodySnapshotFreezer::copy missing body for block " + std::to_string(i)};
-        BlockBodyForStorage& value = *value_opt;
-        // remap to sequential values without gaps (see txnum.go)
-        value.base_txn_id = base_txn_id;
-        base_txn_id += value.txn_count;
-        *out++ = value;
+        BlockBody body;
+        bool found = read_canonical_body(txn, i, /* read_senders = */ true, body);
+        if (!found) throw std::runtime_error{"TransactionSnapshotFreezer::copy missing body for block " + std::to_string(i)};
+
+        *out++ = system_tx;
+        for (auto& value : body.transactions) {
+            *out++ = value;
+        }
+        *out++ = system_tx;
     }
 }
 
-void BodySnapshotFreezer::cleanup(RWTxn& txn, BlockNumRange range) const {
+void TransactionSnapshotFreezer::cleanup(RWTxn& txn, BlockNumRange range) const {
     for (BlockNum i = range.start, count = 1; i < range.end; ++i, ++count) {
         auto hash_opt = read_canonical_header_hash(txn, i);
         if (!hash_opt) continue;
         auto hash = *hash_opt;
 
-        delete_body(txn, hash, i);
+        delete_senders(txn, hash, i);
+
+        auto body_opt = read_canonical_body_for_storage(txn, i);
+        if (body_opt) {
+            auto& body = *body_opt;
+            delete_transactions(txn, body.base_txn_id + 1, body.txn_count - 2);
+        }
 
         if ((count > 10000) && ((count % 10000) == 0)) {
-            log::Debug("BodySnapshotFreezer") << "cleaned up until block " << i;
+            log::Debug("TransactionSnapshotFreezer") << "cleaned up until block " << i;
         }
     }
 }
